\name{Kinship82}
\alias{Kinship82}
\title{Rosenberg-Kim Kinship Terms Partition Data}
\description{
  Partitions of 15 kinship terms given by 85 female undergraduates at
  Rutgers University who were asked to sort the terms into classes
  \dQuote{on the basis of some aspect of meaning}.
}
\usage{data("Kinship82")}
\format{
  A cluster ensemble of 85 hard partitions of the 15 kinship terms.
}
\details{
  \bibcitet{Rosenberg+Kim:1975} describe an experiment where perceived
  similarities of the kinship terms were obtained from six different
  \dQuote{sorting} experiments.  These \dQuote{original} Rosenberg-Kim
  kinship terms data were published in
  \bibcitet{Arabie+Carroll+DeSarbo:1987}, and are also contained in file
  \file{indclus.data} in the shell archive
  \url{https://netlib.org/mds/indclus.shar}. 

  For one of the experiments, partitions of the terms were printed in
  \bibcitet{Rosenberg:1982}.
  Comparison with the original data indicates that
  the partition data have the \dQuote{nephew} and \dQuote{niece} columns
  interchanged, which is corrected in the data set at hand.
}
\source{
  Table 7.1 in \bibcitet{Rosenberg:1982}, with the \dQuote{nephew} and 
  \dQuote{niece} columns interchanged.
}
\references{
  \bibshow{Arabie+Carroll+DeSarbo:1987,
    Rosenberg+Kim:1975,
    Rosenberg:1982}
}
\keyword{datasets}
