\name{Phonemes}
\alias{Phonemes}
\title{Miller-Nicely Consonant Phoneme Confusion Data}
\description{
  Miller-Nicely data on the auditory confusion of 16 consonant
  phonemes.
}
\usage{data("Phonemes")}
\format{
  A symmetric matrix of the misclassification probabilities of 16
  English consonant phonemes.
}
\details{
  \bibcitet{Miller+Nicely:1955}
  obtained the confusions by exposing female
  subjects to a series of syllables consisting of one of the 16
  consonants followed by the vowel \samp{a} under 17 different
  experimental conditions.  The data provided are obtained from
  aggregating the six so-called flat-noise conditions in which only the
  speech-to-noise ratio was varied into a single matrix of
  misclassification frequencies.
}
\source{
  The data set is also contained in file \file{mapclus.data} in the
  shell archive \url{https://netlib.org/mds/mapclus.shar}.
}
\references{
  \bibshow{Miller+Nicely:1955}
}
\keyword{datasets}
