\name{Phonemes}
\alias{Phonemes}
\title{Miller-Nicely Consonant Phoneme Confusion Data}
\description{
  Miller-Nicely data on the auditory confusion of 16 consonant
  phonemes.
}
\usage{data("Phonemes")}
\format{
  A symmetric matrix of the misclassification probabilities of 16
  English consonant phonemes.
}
\details{
  Miller and Nicely (1955) obtained the confusions by exposing female
  subjects to a series of syllables consisting of one of the 16
  consonants followed by the vowel \samp{a} under 17 different
  experimental conditions.  The data provided are obtained from
  aggregating the six so-called flat-noise conditions in which only the
  speech-to-noise ratio was varied into a single matrix of
  misclassification frequencies.
}
\source{
  The data set is also contained in file \file{mapclus.data} in the
  shell archive \url{https://netlib.org/mds/mapclus.shar}.
}
\references{
  G. A. Miller and P. E. Nicely (1955).
  An analysis of perceptual confusions among some English consonants.
  \emph{Journal of the Acoustical Society of America}, \bold{27},
  338--352.
  \doi{10.1121/1.1907526}.
}
\keyword{datasets}
